Health and Hull Display
Displays Enemy Health and Player Health, Work In Progress

+++ Assets/Scripts/Enemy/EnemyAircraftAiInput.cs
@@ -124,7 +124,7 @@ public class EnemyAircraftAiInput : MonoBehaviour
 
     private void CheckHealth()
     {
-        int health = m_health.CurrentHealth;
+        float health = m_health.CurrentHealth;
 
         if (health <= m_fleeHealth 
             && m_playerInRange)
